+++ src/server/game/Handlers/NPCHandler.cpp
@@ -277,10 +277,10 @@ void WorldSession::HandleTrainerBuySpellOpcode(WorldPacket & recv_data)
     uint32 nSpellCost = uint32(floor(trainer_spell->spellCost * _player->GetReputationPriceDiscount(unit)));
 
     // check money requirement
-    if (!_player->HasEnoughMoney(nSpellCost))
+    if (!_player->HasEnoughMoney(uint64(nSpellCost)))
         return;
 
-    _player->ModifyMoney(-int32(nSpellCost));
+    _player->ModifyMoney(-int64(nSpellCost));
 
     unit->SendPlaySpellVisual(179); // 53 SpellCastDirected
     unit->SendPlaySpellImpact(_player->GetGUID(), 362); // 113 EmoteSalute
