+++ src/server/game/Handlers/PetitionsHandler.cpp
@@ -187,7 +187,7 @@ void WorldSession::HandlePetitionBuyOpcode(WorldPacket & recv_data)
         return;
     }
 
-    if (!_player->HasEnoughMoney(cost))
+    if (!_player->HasEnoughMoney(uint64(cost)))
     {                                                       //player hasn't got enough money
         _player->SendBuyError(BUY_ERR_NOT_ENOUGHT_MONEY, creature, charterid, 0);
         return;
